requirements.txt
django==3.2.7
djangorestframework==3.12.4
ipython==7.28.0
ipdb==0.13.9
psycopg2==2.9.1
django-extensions==3.1.3
pytest-django==4.4.0
